cso: Migrate OpenStack presubmits
Migrate the OpenStack presubmits to the a minimal test suite. Also
add `e2e-openstack` to the releases that do not have it.

Make `e2e-openstack-parallel` available starting from release 4.8.

Releases prior to 4.8 only get the parallel suite, because the minimal
suite is not available.

All jobs are on-demand and optional.

Implements [OSASINFRA-2291](https://issues.redhat.com/browse/OSASINFRA-2291)